refactor(smb): use typed enums for SMB2 protocol constants
- Convert Command constants to typed SMB2Command enum with String() method
- Convert Status constants to typed Status enum with String() method
- Convert Flags constants to typed SMB2Flags enum
- Update header package (encoder, header, parser) to use typed enums
- Update all SMB2 handlers to use typed enums for commands and status codes
- Add comprehensive header package tests for encoding/decoding
- Add package documentation for handlers

Typed enums provide IDE autocomplete, compile-time safety, and
better debugging output with String() methods.

+++ internal/protocol/smb/v2/handlers/mock_data.go
@@ -38,7 +38,7 @@ func (h *Handler) GetMockFiles(shareName, path string) map[string]*MockFile {
 				Created:    now,
 				Modified:   now,
 				Accessed:   now,
-				Attributes: types.FileAttributeNormal,
+				Attributes: uint32(types.FileAttributeNormal),
 			},
 			"subdir": {
 				Name:       "subdir",
@@ -47,7 +47,7 @@ func (h *Handler) GetMockFiles(shareName, path string) map[string]*MockFile {
 				Created:    now,
 				Modified:   now,
 				Accessed:   now,
-				Attributes: types.FileAttributeDirectory,
+				Attributes: uint32(types.FileAttributeDirectory),
 			},
 		}
 	}
@@ -70,7 +70,7 @@ func (h *Handler) GetMockFile(shareName, path string) *MockFile {
 			Created:    time.Now(),
 			Modified:   time.Now(),
 			Accessed:   time.Now(),
-			Attributes: types.FileAttributeDirectory,
+			Attributes: uint32(types.FileAttributeDirectory),
 		}
 	}
